added password file to git ignore
this is adding the password file to git ignore and also added decryption

If you want the code I wrote with no research then it is still commented out in these lines.

for full functionality to be cool I added decryption so I can actually use this myself and be more secure.